feat(mascot): persist color selection app-wide via Redux
Read mascotColor from Redux in HumanPage and pass the palette's
primaryColor/secondaryColor as ARGB integers to RiveMascot. Extract
hexToArgbInt into mascotPalette.ts so both HumanPage and MascotPanel
share the same converter. Selecting a swatch in settings now updates
the mascot everywhere instantly.
